+++ R/GVN.R
@@ -3,6 +3,7 @@
 #' Graph equivalent of the Von Neummann variance estimator.
 #'
 #' @export GVN
+#' @importFrom Matrix diag
 #' @param y Noisy data.
 #' @param A Adjacency matrix.
 #' @param L Laplacian matrix.
@@ -21,9 +22,9 @@
 #' @references
 #' von Neumann, J. (1941). Distribution of the ratio of the mean square successive difference to the variance. \emph{Ann. Math. Statistics}, 35(3), 433--451.
 #'
-#' de Loynes, B., Navarro, F., Olivier, B. (2019). Data-driven Thresholding in Denoising with Spectral Graph Wavelet Transform. arXiv preprint arXiv:1906.01882.
+#' de Loynes, B., Navarro, F., Olivier, B. (2021). Data-driven thresholding in denoising with Spectral Graph Wavelet Transform. Journal of Computational and Applied Mathematics, Vol. 389.
 
 GVN <- function(y, A, L) {
-  sig <- 0.5 * sum(A * outer(y, y, "-")^2)/sum(diag(L))
+  sig <- 0.5 * sum(A * outer(y, y, "-")^2)/sum(Matrix::diag(L))
   return(sig)
 }