+++ R/swissroll.R
@@ -6,15 +6,15 @@
 #' \deqn{Sy=\pi (b^2-a^2)y/2}{Sy=pi**(b**2-a**2)*y/2}
 #'
 #' @export swissroll
-#' @import stats
+#' @importFrom stats runif
 #' @param a,b Shape parameters.
 #' @param N Number of points drawn.
 #' @param seed Optionally specify a RNG seed (for reproducible experiments).
 #' @return N x 3 array for 3d points.
 #' @examples
 #' pts <- swissroll(N=500, seed=0, a=1, b=4)
-#' scatterplot3d::scatterplot3d(pts, y=NULL, z=NULL)
 #' @seealso \code{\link{adjacency_mat}}
+
 swissroll <- function(N = 500, seed = NULL, a = 1, b = 4) {
   set.seed(seed)
   x <- runif(N)