+++ inst/doc/gasper_vignette.R
@@ -1,10 +1,9 @@
-## ----setup, include = FALSE----------------------------------------------
+## ----setup, include = FALSE---------------------------------------------------
 library(gasper)
-library(rwavelet)
 if (as.numeric(R.version$minor)>6){
   RNGkind(sample.kind = "Rounding")
 }
-set.seed(0)
+set.seed(434343)
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
@@ -13,46 +12,63 @@ knitr::opts_chunk$set(
   fig.align="center"
 )
 
-## ------------------------------------------------------------------------
-graphname <- "netz4504"
+## -----------------------------------------------------------------------------
+graphname <- "grid1"
 groupname <- "AG-Monien"
-download_graph(graphname,groupname)
-attributes(netz4504)
+download_graph(graphname, groupname)
+attributes(grid1)
 
-## ----fig.show='hold'-----------------------------------------------------
-f <- rnorm(nrow(netz4504$xy))
-plot_graph(netz4504)
-plot_signal(netz4504, f,size = f)
+## -----------------------------------------------------------------------------
+str(grid1$sA)
 
-## ------------------------------------------------------------------------
-A <- full(rlogo$sA)
+## -----------------------------------------------------------------------------
+head(grid1$xy, 3)
+
+## -----------------------------------------------------------------------------
+grid1$dim
+
+## -----------------------------------------------------------------------------
+cat(readLines(grid1$info, n=14), sep = "\n")
+
+## ---- fig.show='hold'---------------------------------------------------------
+f <- rnorm(nrow(grid1$sA))
+plot_graph(grid1)
+plot_signal(grid1, f, size = 2)
+
+## -----------------------------------------------------------------------------
+A <- grid1$sA
 L <- laplacian_mat(A)
-val1 <- eigendec(L)
+val1 <- eigensort(L)
 evalues <- val1$evalues
 evectors <- val1$evectors
+#- largest eigenvalue
 lmax <- max(evalues)
+#- parameter that controls the scale number
 b <- 2
-kmax <- floor(log(lmax)/log(b)) + 2
 tf <- tight_frame(evalues, evectors, b=b)
 
-## ------------------------------------------------------------------------
-x1 <- rlogo$xy[,1]
-x2 <- rlogo$xy[,2]
-n <- length(x1)
+## ----fig.width=5--------------------------------------------------------------
+plot_filter(lmax,b)
+
+## -----------------------------------------------------------------------------
+n <- nrow(L)
 f <- randsignal(0.01, 3, A)
 sigma <- 0.01
 noise <- rnorm(n, sd = sigma)
 tilde_f <- f + noise
 
-## ----fig.show='hold'-----------------------------------------------------
-plot_signal(rlogo, f, size = 3)
-plot_signal(rlogo, tilde_f, size = 3)
+## ----fig.show='hold'----------------------------------------------------------
+plot_signal(grid1, f, size = 2)
+plot_signal(grid1, tilde_f, size = 2)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 wcn <- analysis(tilde_f,tf)
 wcf <- analysis(f,tf)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
+# wcf <- forward_sgwt(f, evalues, evectors, b=b)
+
+## -----------------------------------------------------------------------------
 diagWWt <- colSums(t(tf)^2)
 thresh <- sort(abs(wcn))
 opt_thresh <- SURE_MSEthresh(wcn, 
@@ -64,18 +80,21 @@ opt_thresh <- SURE_MSEthresh(wcn,
                            NA,
                            policy = "dependent")
 
-
-## ------------------------------------------------------------------------
-plot(thresh,opt_thresh$res$MSE,type="l",xlab = "t",ylab = "risk")
-lines(thresh,opt_thresh$res$SURE-n*sigma^2,col="red")
+## -----------------------------------------------------------------------------
+plot(thresh, opt_thresh$res$MSE,
+     type="l", xlab = "t", ylab = "risk")
+lines(thresh, opt_thresh$res$SURE-n*sigma^2, col="red")
 legend("topleft", legend=c("MSE", "SURE"),
        col=c("black", "red"),lty=1)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 hatf_oracle <- synthesis(opt_thresh$wc[,opt_thresh$min[1]], tf)
 hatf_SURE  <- synthesis(opt_thresh$wc[,opt_thresh$min[2]], tf)
 
 SNR(f,tilde_f)
 SNR(f,hatf_oracle)
 SNR(f,hatf_SURE)
+
+## -----------------------------------------------------------------------------
+#hatf_oracle <- inverse_sgwt(opt_thresh$wc[,opt_thresh$min[1]], evalues, evectors, b)